project/inkwell-studio/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
bleach==6.2.0
